Trying to get logout on every page
